mac compat

+++ homedir/.zshrc
@@ -3,9 +3,25 @@ DEFAULT_USER="cole"
 ##############################################################################
 # Environment Detection (used in ~/.zprofile as well)
 ##############################################################################
-export PLATFORM_DISTRO="$(source /etc/os-release; echo "${ID}")"
-export PLATFORM_ARCH="$(echo "amd64")"
-export PLATFORM_OS="$(echo "linux")"
+export HOSTNAME="$(hostname)"
+export UNAMESTR="$(uname)"
+
+export PLATFORM_OS="unknown"
+export PLATFORM_DISTRO="unknown"
+export PLATFORM_ARCH="unknown"
+export NPROC=1
+
+if [[ "${UNAMESTR}" == "Darwin" ]]; then
+	export PLATFORM_OS="mac"
+	export PLATFORM_DISTRO=""
+	export PLATFORM_ARCH="amd64"
+	export NPROC="$(sysctl -n hw.ncpu)"
+elif [[ "${UNAMESTR}" == "Linux" ]]; then
+	export PLATFORM_OS="linux"
+	export PLATFORM_DISTRO="$(source /etc/os-release; echo "${ID}")"
+	export PLATFORM_ARCH="amd64"
+	export NPROC="$(nproc)"
+fi
 
 ##############################################################################
 # config for stuff loaded by zplug
@@ -69,14 +85,30 @@ zplug load
 ###############################################################################
 # coreutils - workaround for termite
 ###############################################################################
-if true; then
-	dircolorstemp="$(mktemp -d)"
-	dircolors -p > "$dircolorstemp/dircolors"
-	last_line="$(grep -n TERM $dircolorstemp/dircolors | tail -1 | cut -d : -f 1)"
-	next_line="$((last_line+1))"
-	sed -i "${next_line}i TERM xterm-termite" "$dircolorstemp/dircolors"
-	eval $(dircolors "$dircolorstemp/dircolors")
-	rm -rf "${dircolorstemp}"
+if [[ "${PLATFORM_OS}" == "linux" ]]; then
+	if true; then
+		dircolorstemp="$(mktemp -d)"
+		dircolors -p > "$dircolorstemp/dircolors"
+		last_line="$(grep -n TERM $dircolorstemp/dircolors | tail -1 | cut -d : -f 1)"
+		next_line="$((last_line+1))"
+		sed -i "${next_line}i TERM xterm-termite" "$dircolorstemp/dircolors"
+		eval $(dircolors "$dircolorstemp/dircolors")
+		rm -rf "${dircolorstemp}"
+	fi
+fi
+
+
+###############################################################################
+# colored ls output
+###############################################################################
+if [[ "${PLATFORM_OS}" == "linux" ]]; then
+	alias l='ls -alh'
+	alias ll='ls -l'
+	alias ls='ls --color=auto'
+elif [[ "${PLATFORM_OS}" == "mac" ]]; then
+	alias l='ls -alh'
+	alias ll='ls -l'
+	alias ls='ls -G'
 fi
